% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.messages.loop_values.R
\name{download.messages.loop_values}
\alias{download.messages.loop_values}
\title{Inside function}
\usage{
download.messages.loop_values(
  devices_toshow,
  indv_loop,
  error_count,
  Tres,
  down_f,
  token,
  type,
  maxrounds,
  show_count
)
}
\arguments{
\item{devices_toshow}{parameter already created inside the above mentioned functions}

\item{indv_loop}{parameter already created inside the above mentioned functions}

\item{Tres}{parameter already created inside the above mentioned functions}

\item{down_f}{parameter already created inside the above mentioned functions}

\item{token}{parameter already created inside the above mentioned functions}

\item{type}{parameter already created inside the above mentioned functions}

\item{maxrounds}{parameter already created inside the above mentioned functions}

\item{show_count}{parameter already created inside the above mentioned functions}

\item{error_coun}{parameter already created inside the above mentioned functions}
}
\value{
updated values for indv_loop, error_count, and Tres  for each round of the loop of download.
}
\description{
This function is only to avoid code repetition inside ecot.whole.download, ecot.update.download and ecot.whole.devices
}
\examples{
download.messages.loop_values(devices_toshow,indv_loop,error_count,Tres)

}
