% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.API.request.R
\name{ecot.API.request}
\alias{ecot.API.request}
\title{Makes a call to ecotopia API to download the indicated data.}
\usage{
ecot.API.request(
  token,
  device_id,
  ndevicelimit = 1000,
  datestart_updates = NA,
  type = "GPS"
)
}
\arguments{
\item{token}{a character string giving your token for an open session. For obtaining it, see ecot.token.}

\item{device_id}{a character string of the device id from the table of devices information obtained with a call to the API. For obtaining it, see ecot.indvs. The 3 different identifiers (uuid, device number, and device id) that appears on the Ecotopia webpage are not useful for this.}

\item{ndevicelimit}{the maximum number of data to download on each request to the API. The maximum available is 1000, which is the default value.}

\item{datestart_updates}{a character string giving a date on the format "Y\%-\%m-\%d \%H:\%M:$S" indicating date from which start the downloading. If this is not provided the downloads will start at first date according to "Deployment date" on Ecotopia webpage.}

\item{type}{a character string indicating the kind of data to download. It can be "GPS", "Env", "ODBA" or "Acc"}
}
\value{
a list of data frame. Each element of the list is the downloadED information for each individual.
}
\description{
ecot.API.request makes an unique request to Ecotopia API based on the chosen type of data, device id and date (optional) for a defined number (optional but always <= 1000) of resulting data.
}
\examples{
\dontrun{
ecot.downloads(ecot.token("abc","passw"), "6267d0fae75c8ef26173d757", type = "Env")
}

}
