% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.downloads.R
\name{ecot.downloads}
\alias{ecot.downloads}
\title{Download individual information based on a date}
\usage{
ecot.downloads(
  token,
  device_id,
  ndevicelimit = 1000,
  type = "GPS",
  maxrounds = NA,
  datestart_updates = NA,
  show_count = F
)
}
\arguments{
\item{token}{a character string giving your token for an open session. For obtaining it, see ecot.token.}

\item{device_id}{a character string of the device id from the table of devices information obtained with the API. For obtaining it, see ecot.indvs. The 3 different identifiers (uuid, device number, and device id) that appears on the Ecotopia web are not useful for this.}

\item{ndevicelimit}{the maximun number of data to download on each request to the API. The maximun available is 1000, which is the default value.}

\item{type}{a character string indicating the kind of data to download. It can be "GPS", "Env", "ODBA" or "Acc". The default value is "GPS".}

\item{maxrounds}{the maximun number of requests to the API. It means that the maximun number of rows donwloaded would be this value multiplied by the value selected on ndevicelimit.}

\item{datestart_updates}{a character string giving a date on the format "Y\%-\%m-\%d \%H:\%M:$S" indicating date from which start the downloading. If this is not provided the downloads will start at first date according to "Deployment date" on Ecotopia webpage.}

\item{show_count}{a logical (TRUE or FALSE) indicating if an indicator of the current data that it is downloading should appears. The indicator is just a string with the last date of each 1e3 rows.}
}
\value{
a list of data frame. Each element of the list is the downloadED information for each individual.
}
\description{
ecot.downloads makes a serie of request to Ecotopia API based on the choosen type of data, device id and date (optional) for a defined number (optional) of resulting data.
}
\examples{
\dontrun{
ecot.downloads(ecot.token("abc","passw"), "6267d0fae75c8ef26173d757", type = "Env")
}

}
