% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.JSON_to_df.R
\name{ecot.JSON_to_df}
\alias{ecot.JSON_to_df}
\title{Transforn lists from JSON files to data.frames}
\usage{
ecot.JSON_to_df(y)
}
\arguments{
\item{y}{a list from a JSON file of GPS/Env/ODBA/Acc data after applying content(x, as = "parsed")}
}
\value{
a list of data frames each one containing information of one device.
}
\description{
Transforn lists resulting from a JSON files to a data frame. It is only for JSON files obtained from a call to the API intended to obtain GPS data from network, ODBA data, Environmental data or accelerometer raw data.
}
\examples{
## not to run, this is used inside ecot.updated.download
Tloop <- ecot.downloads(token = token, indv_id = devID_act[i], type = type, maxrounds = maxrounds, datestart_updates = max_dates[i])
tobind <- ecot.JSON_to_df(Tloop)

}
