% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.token.R
\name{ecot.token}
\alias{ecot.token}
\title{ecot.token}
\usage{
ecot.token(user, psw)
}
\arguments{
\item{user}{a string corresponding to your user name account in Ecotopia}

\item{psw}{a string corresponding to the password of your account in Ecotopia}
}
\value{
It returns a string corresponding to your token for a limited session of API communication
}
\description{
ESTO ES DESCRIP
}
\examples{
ecot.token("abc","passwd") # where your user names is "abc" and your password is "passwd"
}
