% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.unlist.JSON.indvs.R
\name{ecot.unlist.JSON.indvs}
\alias{ecot.unlist.JSON.indvs}
\title{Transforn lists from JSON files to data.frames}
\usage{
ecot.unlist.JSON.indvs(x)
}
\arguments{
\item{x}{a list from a JSON file of devices information after applying content(x, as = "parsed")}
}
\value{
a list of data frames each one containing information of one device.
}
\description{
Transforn lists resulting from a JSON files to a data frame. It is only for JSON files obtained from a call to the API intended to obtain devices information.
}
\examples{
pull <- lapply(input, ecot.unlist.JSON.indvs) # used inside ecot.indv

}
