% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.update.download.R
\name{ecot.update.download}
\alias{ecot.update.download}
\title{Download the whole data base}
\usage{
ecot.update.download(
  user,
  psw,
  token,
  type = "GPS",
  devID,
  max_dates,
  maxrounds = NA,
  show_count = F
)
}
\arguments{
\item{user}{a character string giving to your user name account in Ecotopia. Only neccesary if token is not provided.}

\item{psw}{a character string giving the password of your account in Ecotopia. Only neccesary if token is not provided.}

\item{token}{a character string giving your token for an open session.  Only neccesary if user and psw is not provided. For obtaining it, see ecot.token.}

\item{type}{a character string indicating the kind of data to download. It can be "GPS", "Env", "ODBA" or "Acc"}

\item{devID}{a character vector giving the device id to update. Only those active will be updated. the identifier to use is the one from the table of devices information obtained with a call to the API. For obtaining it, see ecot.indvs. The 3 different identifiers (uuid, device number, and device id) that appears on the Ecotopia webpage are not useful for this.}

\item{max_dates}{a character vector giving the date of the last data alrady download of each device on the format "Y\%-\%m-\%d \%H:\%M:$S". If this is not provided the downloads will start at first date according to "Deployment date" on Ecotopia webpage. This argument admits tenths of a second which is normally the case for acc data.}

\item{maxrounds}{the maximun number of requests to the API. It means that the maximun number of data donwloaded by individual would be this value multiplied by 1000 (maximun available to be downloaded on one request).}

\item{show_count}{a logical (TRUE or FALSE) indicating if an indicator of the current data that it is downloading should appears. The indicator is just a string with the last date of each 1e3 rows.}
}
\value{
a data frame containing all the available information that the Ecotopia API provide for each type of data. Only information from active devices can be downloaded. In the case of accelerometer data, the function returns a list of 2 elements, the first one containing the information for each acc measurement and the second containing al the acc samples corresponding to each measurement.
}
\description{
Download the whole data base
}
\examples{
T_Env <- ecot.whole.download("abc","passw", type = "Env")

}
