% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.whole.devices.R
\name{ecot.whole.devices}
\alias{ecot.whole.devices}
\title{Download the whole data base}
\usage{
ecot.whole.devices(
  user,
  psw,
  token,
  type = "GPS",
  devices,
  maxrounds = NA,
  show_count = F,
  max_devs
)
}
\arguments{
\item{user}{a character string giving to your user name account in Ecotopia. Only neccesary if token is not provided.}

\item{psw}{a character string giving the password of your account in Ecotopia. Only neccesary if token is not provided.}

\item{token}{a character string giving your token for an open session.  Only neccesary if user and psw is not provided. For obtaining it, see ecot.token.}

\item{type}{a character string indicating the kind of data to download. It can be "GPS", "Env", "ODBA" or "Acc"}

\item{devices}{a vector containing the devices number showed on Ecotopia webpage or the UUID to be downloaded.}

\item{maxrounds}{the maximun number of requests to the API. It means that the maximun number of data donwloaded by individual would be this value multiplied by 1000 (maximun available to be downloaded on one request).}

\item{show_count}{a logical (TRUE or FALSE) indicating if an indicator of the current data that it is downloading should appears. The indicator is just a string with the last date of each 1e3 rows.}

\item{max_devs}{the maximun number of active devices to download. Optional.}
}
\value{
a data frame containing all the available information that the Ecotopia API provide for each type of data. Only information from active devices can be downloaded. In the case of accelerometer data, the function returns a list of 2 elements, the first one containing the information for each acc measurement and the second containing al the acc samples corresponding to each measurement.
}
\description{
Download the whole data base
}
\examples{
T_Env <- ecot.whole.download("abc","passw", type = "Env")

}
