% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecot.token.R
\name{ecot.token}
\alias{ecot.token}
\title{Obtain an open session token}
\usage{
ecot.token(user, psw)
}
\arguments{
\item{user}{a character string giving to your user name account in Ecotopia.}

\item{psw}{a character string giving the password of your account in Ecotopia.}
}
\value{
a character string indicating your token for a limited session of API communication.
}
\description{
ecot.token returns an open session token codifiying the your password and making a "post" call to the API. The token has an expiration date but a new one can be requested as many times as needed.
}
\examples{
ecot.token("abc","passwd") # where your user names is "abc" and your password is "passwd"
}
